requirements.txt
# Core Streamlit framework
streamlit>=1.28.0

# PDF/DOCX parsing (approved packages)
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=0.8.11
docx2txt>=0.8

# NLP and text processing (approved packages)
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0

# LangChain ecosystem (approved packages)
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
langgraph>=0.0.20
langsmith>=0.0.60

# Vector stores (approved packages)
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Google Gemini API
google-generativeai>=0.3.0

# Utilities for matching and processing
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.20.0
numpy>=1.24.0
pandas>=2.1.0